add Death dtos validation

+++ src/Models/Dtos/Death/DeathDto.cs
@@ -10,7 +10,7 @@ namespace AirsoftBattlefieldManagementSystemAPI.Models.Dtos.Death
         public int RoomId { get; set; }
         public decimal Longitude { get; set; }
         public decimal Latitude { get; set; }
-        public short Accuracy { get; set; }
+        public decimal Accuracy { get; set; }
         public short Bearing { get; set; }
         public DateTime Time { get; set; }
     }
