+++ src/Models/Dtos/Death/PutDeathDto.cs
@@ -6,7 +6,7 @@ namespace AirsoftBattlefieldManagementSystemAPI.Models.Dtos.Death
     {
         public decimal? Longitude { get; set; }
         public decimal? Latitude { get; set; }
-        public short? Accuracy { get; set; }
+        public decimal? Accuracy { get; set; }
         public short? Bearing { get; set; }
         public DateTime? Time { get; set; }
     }
